bundle.yml: use github as stagingUrl (#99)
Staging s3 bucket/repo is not considered stable, since we often overwrite it with a copy of the production repo. This can cause the downloader to fail, because a prerelease that exists in GitHub is not found in the s3 repo.

With this change, prereleases will be downloaded directly from GitHub releases.